Seperated card screen from main
